A qualitative and quantitative analysis of the COVID-19 pandemic
Paper title: A qualitative and quantitative analysis of the COVID-19 pandemic by a two-side fractional-order compartmental model. Outbreak data from the United States and Brazil were used to demonstrate the effectiveness of the model. The values of initial populations in this work are obtained from the data. We calculate parameters and numerical approximate solutions of the model using Simulink Design Optimization of MATLAB. We can identify the parameters in the model via fractional Adams-Bashforth-Moulton method and nonlinear least squares.